removing google finance tools from project
