--- prosper/datareader/google/__init__.py
@@ -1,1 +1,0 @@
-from . import news
